Original impersonation with glitch
This is the original impersonation script of the english SD card. Unfortunately it has an error when Robi is on his seat. At the end of the routine it will stand up and destroy the power plug!
Never ask "Say something funny" or "Do a mime for me" when Robi is on his seat if this script is still on the card.